+++ src/DataProvider/CampaignScreenCollectionDataProvider.php
@@ -11,12 +11,10 @@ use App\Repository\ScreenCampaignRepository;
 use App\Utils\ValidationUtils;
 use Doctrine\ORM\Tools\Pagination\Paginator as DoctrinePaginator;
 use Symfony\Component\HttpFoundation\RequestStack;
-use Symfony\Component\Security\Core\Security;
 
 final class CampaignScreenCollectionDataProvider implements ContextAwareCollectionDataProviderInterface, RestrictedDataProviderInterface
 {
     public function __construct(
-        private Security $security,
         private RequestStack $requestStack,
         private ScreenCampaignRepository $screenCampaignRepository,
         private ValidationUtils $validationUtils,
@@ -30,26 +28,26 @@ final class CampaignScreenCollectionDataProvider implements ContextAwareCollecti
 
     public function getCollection(string $resourceClass, string $operationName = null, array $context = []): Paginator
     {
-        $itemsPerPage = (int) $this->requestStack->getCurrentRequest()->query->get('itemsPerPage', '10');
-        $page = (int) $this->requestStack->getCurrentRequest()->query->get('page', '1');
-        $id = $this->requestStack->getCurrentRequest()->attributes->get('id');
+        $currentRequest = $this->requestStack->getCurrentRequest();
+        $itemsPerPage = $currentRequest->query?->get('itemsPerPage') ?? 10;
+        $page = $currentRequest->query?->get('page') ?? 1;
+        $id = $currentRequest->attributes?->get('id') ?? '';
+
         $queryNameGenerator = new QueryNameGenerator();
         $campaignId = $this->validationUtils->validateUlid($id);
 
         // Get playlist to check shared-with-tenants
         $queryBuilder = $this->screenCampaignRepository->getScreensBasedOnCampaign($campaignId);
 
+        // Filter the query-builder with tenant extension.
         foreach ($this->collectionExtensions as $extension) {
             $extension->applyToCollection($queryBuilder, $queryNameGenerator, $resourceClass, $operationName, $context);
-            if ($extension instanceof QueryResultItemExtensionInterface && $extension->supportsResult($resourceClass, $operationName, $context)) {
-                return $extension->getResult($queryBuilder, $resourceClass, $operationName, $context);
-            }
         }
 
-        $firstResult = ($page - 1) * $itemsPerPage;
+        $firstResult = ((int) $page - 1) * (int) $itemsPerPage;
         $query = $queryBuilder->getQuery()
             ->setFirstResult($firstResult)
-            ->setMaxResults($itemsPerPage);
+            ->setMaxResults((int) $itemsPerPage);
 
         $doctrinePaginator = new DoctrinePaginator($query);